+++ src/services/mod.rs
@@ -1,3 +1,4 @@
+pub mod content_cache;
 pub mod content_pipeline;
 pub mod content_provider;
 pub mod device_registry;
@@ -6,7 +7,8 @@ pub mod renderer;
 pub mod template_service;
 pub mod url_signer;
 
-pub use content_pipeline::ContentPipeline;
+pub use content_cache::{CachedContent, ContentCache};
+pub use content_pipeline::{ContentPipeline, DeviceContext};
 pub use content_provider::ContentProvider;
 pub use device_registry::{DeviceRegistry, InMemoryRegistry};
 pub use lua_runtime::{LuaRuntime, ScriptError};